# Cache verification of certificate signatures