--- a/crates/sui-core/src/authority.rs
+++ b/crates/sui-core/src/authority.rs
@@ -92,6 +92,7 @@ use sui_types::{
     SUI_FRAMEWORK_ADDRESS,
 };
 
+pub use crate::authority::authority_per_epoch_store::VerifiedCertificateCache;
 use crate::authority::authority_per_epoch_store::{
     AuthorityPerEpochStore, EpochStartConfiguration,
 };