consensus_validator and authority_server can avoid redundantly verifying the same cert twice.

Using simulated load tests (which provide a gross accounting of all the CPU time spent in a cluster) we see:

With optimization:

    ; SIM_STRESS_TEST_QPS=25 cargo simtest test_simulated_load_basic
        Starting 1 tests across 3 binaries (6 skipped)
            PASS [  20.395s] sui-benchmark::simtest test::test_simulated_load_basic
    ------------
         Summary [  20.395s] 1 tests run: 1 passed, 6 skipped

Without:

    ; SIM_STRESS_TEST_QPS=25 cargo simtest test_simulated_load_basic
        Starting 1 tests across 3 binaries (6 skipped)
            PASS [  29.491s] sui-benchmark::simtest test::test_simulated_load_basic
    ------------
         Summary [  29.491s] 1 tests run: 1 passed, 6 skipped